14 Added Implementation for self-trained model
