+++ app/src/main/java/de/hsaugsburg/teamulster/sohappy/analyzer/BitmapEditor.kt
@@ -1,6 +1,8 @@
 package de.hsaugsburg.teamulster.sohappy.analyzer
 
 import android.graphics.*
+import java.nio.ByteBuffer
+import java.nio.ByteOrder
 
 /**
  * Provides a toolbox for editing a bitmap, like cropping, rotating, resizing and grey scaling.
@@ -53,5 +55,25 @@ object BitmapEditor {
         paint.colorFilter = colorMatrixFilter
         canvas.drawBitmap(copiedBitmap, 0F, 0F, paint)
         return bitmapGreyScale
+    }
+
+    /**
+     * This private function converts a given Bitmap into a ByteBuffer.
+     *
+     * @param bitmap a Bitmap which will be converted into a ByteBuffer
+     * @return ByteBuffer?
+     * */
+    fun convertToByteBuffer(bitmap: Bitmap): ByteBuffer? {
+        val width = bitmap.width
+        val height = bitmap.height
+        val mImgData: ByteBuffer = ByteBuffer
+            .allocateDirect(4 * width * height)
+        mImgData.order(ByteOrder.nativeOrder())
+        val pixels = IntArray(width * height)
+        bitmap.getPixels(pixels, 0, width, 0, 0, width, height)
+        for (pixel in pixels) {
+            mImgData.putFloat(Color.red(pixel).toFloat())
+        }
+        return mImgData
     }
 }
